feat: Add Stake Info component
