+++ apps/web/src/views/LiquidStaking/Stake/index.tsx
@@ -1,18 +1,12 @@
-import { CardBody, Text, Select, RowBetween, Button } from '@pancakeswap/uikit'
+import { CardBody, Text, Select, Button } from '@pancakeswap/uikit'
 import NextLink from 'next/link'
 import { AppHeader } from 'components/App'
 import { useActiveChainId } from 'hooks/useActiveChainId'
 import { useState, useCallback, useEffect, useMemo } from 'react'
 import { useTranslation } from '@pancakeswap/localization'
-import BigNumber from 'bignumber.js'
-import { BIG_ZERO } from '@pancakeswap/utils/bigNumber'
-import { getFullDisplayBalance } from '@pancakeswap/utils/formatBalance'
 import { useLiquidStakingList } from 'views/LiquidStaking/hooks/useLiquidStakingList'
 import { LiquidStakingList } from 'views/LiquidStaking/constants/types'
-import { useExchangeRate } from 'views/LiquidStaking/hooks/useExchangeRate'
-
-import { ExchangeRateTitle } from '../components/ExchangeRateTitle'
-import { LiquidStakingApr } from '../components/LiquidStakingApr'
+import StakeInfo from 'views/LiquidStaking/components/StakeInfo'
 
 interface OptionProps extends LiquidStakingList {
   label: string
@@ -48,20 +42,11 @@ export function LiquidStakingPageStake() {
 
   const handleSortOptionChange = useCallback((option) => setSelectedList(option), [])
 
-  const { exchangeRateList } = useExchangeRate({ decimals: selectedList?.token0?.decimals })
-
   useEffect(() => {
     if (initState) {
       setSelectedList(initState)
     }
   }, [chainId, initState])
-
-  const exchangeRateAmount = useMemo(() => {
-    const pickedRate = exchangeRateList?.find(
-      (i) => i?.contract?.toLowerCase() === selectedList?.contract?.toLowerCase(),
-    )?.exchangeRate
-    return new BigNumber(pickedRate) ?? BIG_ZERO
-  }, [exchangeRateList, selectedList?.contract])
 
   return (
     <>
@@ -76,24 +61,8 @@ export function LiquidStakingPageStake() {
           {t('Choose a pair to liquid stake')}
         </Text>
         {optionsList.length > 0 && <Select mb="24px" options={optionsList} onOptionChange={handleSortOptionChange} />}
-        <RowBetween mb="8px">
-          <ExchangeRateTitle tokenOSymbol={selectedList?.token0?.symbol} token1Symbol={selectedList?.token1?.symbol} />
-
-          {exchangeRateAmount ? (
-            <Text>
-              {t('1 %token0% = %exchangeRateAmount% %token1%', {
-                token0: selectedList?.token0?.symbol,
-                token1: selectedList?.token1?.symbol,
-                exchangeRateAmount: getFullDisplayBalance(exchangeRateAmount, 0, 6),
-              })}
-            </Text>
-          ) : (
-            '-'
-          )}
-        </RowBetween>
-
-        <LiquidStakingApr contract={selectedList?.contract} tokenOSymbol={selectedList?.token0?.symbol} />
-        <NextLink href={`/liquid-staking/${selectedList?.symbol}`}>
+        <StakeInfo selectedList={selectedList} />
+        <NextLink href={`/liquid-staking/${selectedList?.symbol}?contract=${selectedList.contract}`}>
           <Button width="100%">{t('Proceed')}</Button>
         </NextLink>
       </CardBody>
