+++ apps/web/src/views/LiquidStaking/components/StakeInfo/ExchangeRateTitle.tsx
